refactor(List): move exact responsibility from list group to list items
BREAKING CHANGE: ListGroups no longer have the `exact` input property. Instead, the ListItem
should indicate whether it should be active when the exact route is found, or a subset of it. If
ListGroups have active items, then they are automatically opened when the web page is loaded.
Checking is now done via the `active` property of the ListItem component.