refactor(Menu): improve menu
BREAKING CHANGE: The menu items are no longer passed in the `default` slot but in the `items` slot.
The activator is now passed in the `default` slot instead of the `activator` slot.
BREAKING CHANGE: The menu component has been rewritten, therefore some input properties and the
settings token have changed.